+++ src/common/dependency-manager/local-manager.ts
@@ -4,8 +4,8 @@ import fs from 'fs-extra';
 import path from 'path';
 import DependencyManager, { DependencyNode, EnvironmentConfig, EnvironmentConfigBuilder, Refs } from '../../dependency-manager/src';
 import DependencyGraph from '../../dependency-manager/src/graph';
-import { ComponentConfig } from '../../dependency-manager/src/spec/component/base';
-import { ComponentConfigBuilder } from '../../dependency-manager/src/spec/component/builder';
+import { ComponentConfigBuilder } from '../../dependency-manager/src/spec/component/component-builder';
+import { ComponentConfig } from '../../dependency-manager/src/spec/component/component-config';
 import { Dictionary } from '../../dependency-manager/src/utils/dictionary';
 import { flattenValidationErrorsWithLineNumbers, ValidationErrors } from '../../dependency-manager/src/utils/errors';
 import { readIfFile } from '../../dependency-manager/src/utils/files';
