+++ src/dependency-manager/src/index.ts
@@ -5,16 +5,19 @@ export * from './graph/node';
 export * from './graph/node/service';
 export * from './graph/node/task';
 export * from './spec/base-spec';
+export * from './spec/common/build-spec';
+export * from './spec/common/deploy-spec';
 export * from './spec/common/interface-spec';
 export * from './spec/common/liveness-probe-spec';
 export * from './spec/common/parameter-spec';
-export * from './spec/component/base';
-export * from './spec/component/builder';
-export * from './spec/environment/base';
-export * from './spec/environment/builder';
-export * from './spec/resource/base';
-export * from './spec/service/base';
-export * from './spec/task/base';
+export * from './spec/common/volume-spec';
+export * from './spec/component/component-builder';
+export * from './spec/component/component-config';
+export * from './spec/environment/environment-builder';
+export * from './spec/environment/environment-config';
+export * from './spec/resource/resource-config';
+export * from './spec/service/service-config';
+export * from './spec/task/task-config';
 export * from './utils/refs';
 export * from './utils/slugs';
