+++ src/dependency-manager/src/manager.ts
@@ -8,10 +8,10 @@ import GatewayNode from './graph/node/gateway';
 import InterfacesNode from './graph/node/interfaces';
 import { TaskNode } from './graph/node/task';
 import { InterfaceSpec } from './spec/common/interface-spec';
-import { ComponentConfig } from './spec/component/base';
-import { ComponentConfigBuilder } from './spec/component/builder';
-import { EnvironmentConfig } from './spec/environment/base';
-import { EnvironmentConfigBuilder } from './spec/environment/builder';
+import { ComponentConfigBuilder } from './spec/component/component-builder';
+import { ComponentConfig } from './spec/component/component-config';
+import { EnvironmentConfigBuilder } from './spec/environment/environment-builder';
+import { EnvironmentConfig } from './spec/environment/environment-config';
 import { Dictionary } from './utils/dictionary';
 import { flattenValidationErrors, ValidationErrors } from './utils/errors';
 import { escapeJSON, interpolateString, normalizeInterpolation, prefixExpressions, removePrefixForExpressions, replaceBrackets } from './utils/interpolation';
