+++ src/dependency-manager/src/spec/common/build-spec.ts
@@ -1,0 +1,7 @@
+import { Dictionary } from '../../utils/dictionary';
+
+export interface BuildSpec {
+  context?: string;
+  args?: Dictionary<string>;
+  dockerfile?: string;
+}
