+++ src/dependency-manager/src/spec/common/deploy-spec.ts
@@ -1,0 +1,11 @@
+import { Dictionary } from '../../utils/dictionary';
+
+export interface DeployModuleSpec {
+  path: string;
+  inputs: Dictionary<string>;
+}
+
+export interface DeploySpec {
+  strategy: string;
+  modules: Dictionary<DeployModuleSpec>;
+}
