+++ src/dependency-manager/src/spec/common/volume-spec.ts
@@ -1,0 +1,7 @@
+
+export interface VolumeSpec {
+  mount_path?: string;
+  host_path?: string;
+  description?: string;
+  readonly?: boolean;
+}
