+++ src/dependency-manager/src/spec/component/component-builder.ts
@@ -7,9 +7,9 @@ import path from 'path';
 import { Dictionary } from '../../utils/dictionary';
 import { flattenValidationErrorsWithLineNumbers, ValidationErrors } from '../../utils/errors';
 import { insertFileDataFromRefs } from '../../utils/files';
-import { DeploySpec } from '../resource/base';
-import { ComponentConfig } from './base';
-import { ComponentConfigV1 } from './v1';
+import { DeploySpec } from '../common/deploy-spec';
+import { ComponentConfig } from './component-config';
+import { ComponentConfigV1 } from './component-v1';
 
 class MissingConfigFileError extends Error {
   constructor(filepath: string) {
